requirements.txt
flask
pymongo
fyers_apiv3
requests
APScheduler
python-dotenv
waitress
Werkzeug
pandas
numpy
pytz
kiteconnect
yfinance
Authlib
requests